+++ frontend/src/components/AnchorLinks/index.tsx
@@ -1,0 +1,7 @@
+export { AnchorLinks } from './AnchorLinks';
+
+export type {
+    AnchorLinkItem,
+    AnchorLinksProps,
+    ControlledLink,
+} from './AnchorLinks';
