+++ frontend/src/components/DateInput/index.tsx
@@ -1,0 +1,2 @@
+export { DateInput } from './DateInput';
+export type { DateInputProps } from './DateInput';
