+++ frontend/src/components/Dropdown/index.tsx
@@ -1,0 +1,2 @@
+export { Dropdown } from './Dropdown';
+export type { DropdownProps, DropdownOption } from './Dropdown';
