+++ frontend/src/components/FileUpload/index.tsx
@@ -1,0 +1,2 @@
+export { FileUpload } from './FileUpload';
+export type { FileUploadProps } from './FileUpload';
