+++ frontend/src/components/FormContainer/index.tsx
@@ -1,0 +1,2 @@
+export { FormContainer } from './FormContainer';
+export type { FormContainerProps } from './FormContainer';
