+++ frontend/src/components/InforCard/InfoCard.tsx
@@ -1,21 +1,28 @@
 import { ReactNode } from 'react';
 import { IoClose } from 'react-icons/io5';
 
-interface InfoCardProps {
+export interface InfoCardProps {
     icon?: ReactNode;
     text: string;
     subtext?: string;
     onRemove?: () => void;
 }
 
-const InfoCard: React.FC<InfoCardProps> = ({ icon, text, subtext, onRemove }) => {
+const InfoCard: React.FC<InfoCardProps> = ({
+    icon,
+    text,
+    subtext,
+    onRemove,
+}) => {
     return (
         <div className="flex items-center gap-3 bg-white rounded-full px-4 py-2 shadow-sm">
             {icon && <div className="flex-shrink-0">{icon}</div>}
             <div className="flex-grow">
                 <span className="font-medium">{text}</span>
                 {subtext && (
-                    <span className="text-gray-400 ml-2 text-sm">{subtext}</span>
+                    <span className="text-gray-400 ml-2 text-sm">
+                        {subtext}
+                    </span>
                 )}
             </div>
             {onRemove && (
@@ -31,4 +38,5 @@ const InfoCard: React.FC<InfoCardProps> = ({ icon, text, subtext, onRemove }) =>
     );
 };
 
-export { InfoCard }; 
+export { InfoCard };
+
