+++ frontend/src/components/InforCard/index.tsx
@@ -1,0 +1,2 @@
+export { InfoCard } from './InfoCard';
+export type { InfoCardProps } from './InfoCard';
