+++ frontend/src/components/NotificationBanner/index.tsx
@@ -1,0 +1,2 @@
+export { NotificationBanner } from './NotificationBanner';
+export type { NotificationBannerProps } from './NotificationBanner';
