+++ frontend/src/components/ScrollLink/index.tsx
@@ -1,0 +1,2 @@
+export { ScrollLink } from './ScrollLink';
+export type { ScrollLinkProps } from './ScrollLink';
