+++ frontend/src/components/SocialLinks/index.tsx
@@ -1,0 +1,2 @@
+export { SocialLinks } from './SocialLinks';
+export type { SocialLinksProps } from './SocialLinks';
