+++ frontend/src/components/TeamMembers/index.tsx
@@ -1,0 +1,2 @@
+export { TeamMembers } from './TeamMembers';
+export type { TeamMembersProps } from './TeamMembers';
