+++ frontend/src/components/TextArea/index.tsx
@@ -1,0 +1,2 @@
+export { TextArea } from './TextArea';
+export type { TextAreaProps } from './TextArea';
