+++ frontend/src/components/TextInput/index.tsx
@@ -1,0 +1,2 @@
+export { TextInput } from './TextInput';
+export type { TextInputProps } from './TextInput';
