+++ frontend/src/components/index.ts
@@ -1,26 +1,41 @@
 export { TextInput } from './TextInput';
-export { Dropdown } from './Dropdown';
-export { FileUpload } from './FileUpload';
-export { InfoCard } from './InfoCard';
-export { Button } from './Button';
-export { NotificationBanner } from './NotificationBanner';
-export { FormContainer } from './FormContainer';
-export { ScrollLink } from './ScrollLink';
-export { AnchorLinks } from './AnchorLinks';
-export { TextArea } from './TextArea';
-export { TeamMembers } from './TeamMembers';
-export { SocialLinks } from './SocialLinks';
-export { DateInput } from './DateInput';
-export {
-    PageLayout,
-    Section,
-    Grid,
-    Footer,
-    Header,
-    DashboardTemplate,
-    AdminDashboard,
-    UserDashboard,
-} from './layout';
+export type { TextInputProps } from './TextInput';
 
-import type { AnchorLinkItem } from './AnchorLinks';
-export type { AnchorLinkItem };
+export { Dropdown } from './Dropdown';
+export type { DropdownOption, DropdownProps } from './Dropdown';
+
+export { FileUpload } from './FileUpload';
+export type { FileUploadProps } from './FileUpload';
+
+export { InfoCard } from './InforCard';
+export type { InfoCardProps } from './InforCard';
+
+export { Button } from './Button';
+
+export { NotificationBanner } from './NotificationBanner';
+export type { NotificationBannerProps } from './NotificationBanner';
+
+export { FormContainer } from './FormContainer';
+export type { FormContainerProps } from './FormContainer';
+
+export { ScrollLink } from './ScrollLink';
+export type { ScrollLinkProps } from './ScrollLink';
+
+export { AnchorLinks } from './AnchorLinks';
+export type {
+    AnchorLinksProps,
+    AnchorLinkItem,
+    ControlledLink,
+} from './AnchorLinks';
+
+export { TextArea } from './TextArea';
+export type { TextAreaProps } from './TextArea';
+
+export { TeamMembers } from './TeamMembers';
+export type { TeamMembersProps } from './TeamMembers';
+
+export { SocialLinks } from './SocialLinks';
+export type { SocialLinksProps } from './SocialLinks';
+
+export { DateInput } from './DateInput';
+export type { DateInputProps } from './DateInput';
